Update config.fish adding compatibility with zellij

+++ .config/fish/config.fish
@@ -1,21 +1,29 @@
-set fish_greeting
+if status is-interactive; and not set -q VSCODE_RESOLVING_ENVIRONMENT
+	set fish_greeting
 
-set -x MANPAGER "sh -c 'col -bx | bat -l man -p'"
-set -x MANROFFOPT "-c"
+	set -x MANPAGER "sh -c 'col -bx | bat -l man -p'"
+	set -x MANROFFOPT "-c"
 
-alias ls='exa --color=always --group-directories-first'
-alias la='exa -la --color=always --group-directories-first'
-alias ll='exa -l --color=always --group-directories-first'
-alias lt='exa -aT --color=always --group-directories-first'
+	alias ls='eza --color=always --group-directories-first'
+	alias la='ls -la'
+	alias ll='ls -l'
+	alias lt='ls -aT'
 
-alias grep='grep --color=auto'
+	alias grep='grep --color=auto'
 
-alias ip='ip --color=auto'
+	alias ip='ip --color=auto'
 
-alias cp='cp -i'
-alias mv='mv -i'
-alias rm='rm -i'
+	alias cp='cp -i'
+	alias mv='mv -i'
+	alias rm='rm -i'
 
-alias cat='bat --style snip --theme OneHalfDark'
+	alias cat='bat --style snip --theme OneHalfDark'
 
-starship init fish | source
+	if test "$TERM_PROGRAM" != "vscode"
+		set ZELLIJ_AUTO_ATTACH true
+		set ZELLIJ_AUTO_EXIT true
+		eval (zellij setup --generate-auto-start fish | string collect)
+	end
+
+	starship init fish | source
+end
